+++ .github/workflows/main.yml
@@ -4,9 +4,9 @@ name: CI
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
-    branches: [ main , dev , image_captioning]
+    branches: [ main , dev]
   pull_request:
-    branches: [ dev, main , image_captioning]
+    branches: [ dev, main]
 jobs:
   docker-job:
     strategy:
